improve loading experience
- blocking loading at the component level while results are processed was causing a jarring experience on page load
- hoists processing of results into a shared service relied upon by the routes
- removes layout provider component pattern
- removes fades which also causes some loading jank
