+++ app/templates/components/main-layout.hbs
@@ -1,4 +1,4 @@
-<div class="main ml-5 fade-in {{if this.displayMobileToc "mobile" "desktop"}}">
+<div class="main ml-5 {{if this.displayMobileToc "mobile" "desktop"}}">
   {{#if @sortedGroupedResults}}
     <h1 data-test-deprecations-added-in>Deprecations Added in {{@project}} {{@version}}</h1>
     <div class="intro">
