+++ app/templates/index.hbs
@@ -1,4 +1,4 @@
-<div id="content" class="main fade-in" role="main">
+<div id="content" class="main" role="main">
   <div class="intro">
     <h1 class="anchorable-toc" id="toc_deprecations">Deprecations</h1>
     <p>Periodically, various APIs in Ember.js may be deprecated. During a minor
